+++ CHANGELOG.md
@@ -56,6 +56,7 @@ The present file will list all changes made to the project; according to the
 - `Transfer::transferDropdownNetpoint()` has been renamed to `Transfer::transferDropdownSocket()`.
 
 #### Deprecated
+- Usage of `MyISAM` engine in database, in favor of `InnoDB` engine.
 - Usage of `utf8mb3` charset/collation in database in favor of `utf8mb4` charset/collation.
 - Handling of encoded/escaped value in `autoName()`
 - `Netpoint` has been deprecated and replaced by `Socket`